chore: merge all weather files into MapWeather.h and MapWeather.cpp
