--- src/server/game/World/Weather/Weather.h
@@ -1,100 +1,0 @@
-/*
- * This file is part of the AzerothCore Project. See AUTHORS file for Copyright information
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation; either version 3 of the License, or (at your
- * option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-/// \addtogroup world
-/// @{
-/// \file
-
-#ifndef __WEATHER_H
-#define __WEATHER_H
-
-#include "Common.h"
-#include "SharedDefines.h"
-#include "Timer.h"
-
-class Player;
-
-#define WEATHER_SEASONS 4
-struct WeatherSeasonChances
-{
-    uint32 rainChance;
-    uint32 snowChance;
-    uint32 stormChance;
-};
-
-struct WeatherData
-{
-    WeatherSeasonChances data[WEATHER_SEASONS];
-    uint32 ScriptId;
-};
-
-enum WeatherState : uint32
-{
-    WEATHER_STATE_FINE              = 0,
-    WEATHER_STATE_FOG               = 1,
-    WEATHER_STATE_LIGHT_RAIN        = 3,
-    WEATHER_STATE_MEDIUM_RAIN       = 4,
-    WEATHER_STATE_HEAVY_RAIN        = 5,
-    WEATHER_STATE_LIGHT_SNOW        = 6,
-    WEATHER_STATE_MEDIUM_SNOW       = 7,
-    WEATHER_STATE_HEAVY_SNOW        = 8,
-    WEATHER_STATE_LIGHT_SANDSTORM   = 22,
-    WEATHER_STATE_MEDIUM_SANDSTORM  = 41,
-    WEATHER_STATE_HEAVY_SANDSTORM   = 42,
-    WEATHER_STATE_THUNDERS          = 86,
-    WEATHER_STATE_BLACKRAIN         = 90,
-    WEATHER_STATE_BLACKSNOW         = 106
-};
-
-enum WEATHER_TYPE {
-  WEATHER_TYPE_FINE   = 0x0,
-  WEATHER_TYPE_RAIN   = 0x1,
-  WEATHER_TYPE_SNOW   = 0x2,
-  WEATHER_TYPE_STORM  = 0x3,
-  MAX_WEATHER_TYPE    = 0x4,
-  //TODO:
-  WEATHER_TYPE_THUNDERS   = 86,
-  WEATHER_TYPE_BLACKRAIN  = 90
-};
-
-/// Weather for one zone
-class Weather
-{
-public:
-    Weather(uint32 zone, WeatherData const* weatherChances);
-    ~Weather() = default;
-
-    bool Update(uint32 diff);
-    bool ReGenerate();
-    bool UpdateWeather();
-
-    void SendWeatherUpdateToPlayer(Player* player);
-    void SetWeather(WEATHER_TYPE type, float grade);
-
-    /// For which zone is this weather?
-    [[nodiscard]] uint32 GetZone() const { return m_zone; };
-    [[nodiscard]] uint32 GetScriptId() const { return m_weatherChances->ScriptId; }
-
-private:
-    [[nodiscard]] WeatherState GetWeatherState() const;
-    uint32 m_zone;
-    WEATHER_TYPE m_type;
-    float m_grade;
-    IntervalTimer m_timer;
-    WeatherData const* m_weatherChances;
-};
-#endif
