--- src/server/game/World/Weather/WeatherMgr.h
@@ -1,43 +1,0 @@
-/*
- * This file is part of the AzerothCore Project. See AUTHORS file for Copyright information
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation; either version 3 of the License, or (at your
- * option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-/// \addtogroup world
-/// @{
-/// \file
-
-#ifndef __WEATHERMGR_H
-#define __WEATHERMGR_H
-
-#include "Define.h"
-
-class Weather;
-class Player;
-
-namespace WeatherMgr
-{
-    void LoadWeatherData();
-
-    Weather* FindWeather(uint32 id);
-    Weather* AddWeather(uint32 zone_id);
-    void RemoveWeather(uint32 zone_id);
-
-    void SendFineWeatherUpdateToPlayer(Player* player);
-
-    void Update(uint32 diff);
-}
-
-#endif
