+++ src/server/scripts/Commands/cs_misc.cpp
@@ -43,7 +43,7 @@
 #include "SpellAuras.h"
 #include "TargetedMovementGenerator.h"
 #include "Tokenize.h"
-#include "WeatherMgr.h"
+
 
 /// @todo: this import is not necessary for compilation and marked as unused by the IDE
 //  however, for some reasons removing it would cause a damn linking issue
